less code in __init__.py
